requirements.txt
# Core orchestration
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0

# LLM
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
openai>=1.0.0

# Tool integrations
google-search-results>=2.4.2    # SerpAPI client (imports as `serpapi`)
firecrawl-py>=4.0.0
googlemaps>=4.10.0              # Distance Matrix, Places, Geocoding

# Observability (auto-enabled when LANGCHAIN_TRACING_V2=true)
langsmith>=0.1.0

# Env loading
python-dotenv>=1.0.0

# Graph visualization (used by scripts/print_graph.py to render ASCII diagrams)
grandalf>=0.8
json-repair
